requirements.txt
aiogram==3.1.1
aioschedule @ https://github.com/AleksHeller/python-aioschedule/archive/refs/heads/master.zip
python-dotenv==1.0.0
